made some changes according to comments
