+++ tools/src/wyvern/tools/tests/suites/AntRegressionTestSuite.java
@@ -27,6 +27,7 @@ import wyvern.tools.tests.ReflectionTests;
 import wyvern.tools.tests.RossettaCodeTests;
 import wyvern.tools.tests.StdlibTests;
 import wyvern.tools.tests.TransformTests;
+import wyvern.tools.tests.EffectSeparationTests;
 
 /**
  * This test suite includes all working tests (as of this writing) that
@@ -62,7 +63,8 @@ import wyvern.tools.tests.TransformTests;
             PolymorphicEffectTests.class,
             EffectAnnotationTests.class,
             ModuleSystemTests.class,        // tests the new IL
-            HigherOrderTests.class
+            HigherOrderTests.class,
+            EffectSeparationTests.class
         }
         )
 public class AntRegressionTestSuite { }
